requirements.txt
# Textile QC Web Application - Dependencies
# Python 3.11+

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# Image Processing
numpy>=1.26.0,<2.0
opencv-python-headless>=4.8.0
Pillow>=10.1.0
scikit-image>=0.22.0

# Scientific Computing
scipy>=1.11.0
pandas>=2.1.0
PyWavelets>=1.5.0

# PDF Generation
reportlab>=4.0.0

# Visualization
matplotlib>=3.8.0

# Utilities
python-dateutil>=2.8.0
